Introduced the BasicClass, moving extend() from utils.
