+++ src/utils.js
@@ -15,38 +15,6 @@
 CKEDITOR.define( [ 'utils-lodash', 'lib/lodash/lodash-ckeditor' ], function( lodashIncludes, lodash ) {
 	var utils = {
 		/**
-		 * A mixin function to be used to implement the static `extend()` method in classes. It allows for easy creation
-		 * of subclasses.
-		 *
-		 * @param {Object} [proto] Extensions to be added to the subclass prototype.
-		 * @param {Object} [statics] Additional static properties to be added to the subclass constructor.
-		 * @returns {Object} When executed as a static method of a class, it returns the new subclass constructor.
-		 */
-		extendMixin: function( proto, statics ) {
-			var that = this;
-			var child = ( proto && proto.hasOwnProperty( 'constructor' ) ) ?
-					proto.constructor :
-					function() {
-						that.apply( this, arguments );
-					};
-
-			// Copy the statics.
-			utils.extend( child, this, statics );
-
-			// Use the same prototype.
-			child.prototype = Object.create( this.prototype );
-
-			// Add the new prototype stuff.
-			if ( proto ) {
-				proto = utils.clone( proto );
-				delete proto.constructor;
-				utils.extend( child.prototype, proto );
-			}
-
-			return child;
-		},
-
-		/**
 		 * Creates a spy function (ala Sinon.js) that can be used to inspect call to it.
 		 *
 		 * The following are the present features:
@@ -85,17 +53,3 @@ CKEDITOR.define( [ 'utils-lodash', 'lib/lodash/lodash-ckeditor' ], function( lod
 
 	return utils;
 } );
-
-/**
- * Creates a subclass constructor based on this class.
- *
- * @member utils
- * @method extend
- * @abstract
- * @static
- * @inheritable
- *
- * @param {Object} [proto] Extensions to be added to the subclass prototype.
- * @param {Object} [statics] Extension to be added as static members of the subclass constructor.
- * @returns {Object} The subclass constructor.
- */
